requirements.txt
streamlit
pymupdf
python-docx
google-generativeai
